backend/requirements.txt
requests
beautifulsoup4
PyMuPDF
cohere
qdrant-client
python-dotenv
langchain
langchain_text_splitters
lxml
fastapi
uvicorn
pydantic
google-generativeai